backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.2
python-decouple>=3.8
django-cors-headers>=4.6
gunicorn>=23.0
